Fixed importFiles (#69)
* Whatever

* Oops

* Crap

+++ code/kdTree/parallelApproved/main.cpp
@@ -116,8 +116,12 @@ int main(int argc, char *argv[])
 	int rows = 0, cols = _ROW_WIDTH_;
 
 	// Read data files in
-	int arrayLimit = (FilenameArray.size() * maxRows * cols + 8);
+	unsigned long int arrayLimit = (FilenameArray.size() * maxRows * cols + 8);
 	double *array = new double[arrayLimit];
+
+	// void importFiles(vector<string> files, int myRank,
+	// 	double *myData, int *rows, int *cols, int maxRows,
+	// 	int maxRowsPerFile, unsigned long int arrayLimit)
 
 	importFiles(FilenameArray, myRank, array, &rows, &cols, maxRows, arrayLimit);
